trunk/ (mint@mint) Fixing 'int' type (changin it to 'intmax_t') in the 'cvalue_t.c' file. This now fixes all 'list*.sn;' tests.
git-svn-id: svn+ssh://saulius-grazulis.lt/home/saulius/svn-repositories/compilers/trunk@2971 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ sc/cvalue_t.c
@@ -106,7 +106,7 @@ const_value_t make_const_value( cexception_t *ex,
 
     switch( value_type ) {
     case VT_INT:
-	r.value.i = va_arg( v, int );
+	r.value.i = va_arg( v, intmax_t );
 	break;
     case VT_FLOAT:
 	r.value.f = va_arg( v, double );
